MAINT: rename module scikit_radon -> skimage_radon
